rename: rename schemas to hooks folder
