--- src/schemas/index.ts
@@ -1,1 +1,0 @@
-export * from './profileSchema.ts';
